# backend/core/docs/swagger.yaml
basePath: /api/v1
definitions:
  articles.ArticleInput:
    properties:
      authorname:
        type: string
      content:
        type: string
      draft:
        type: boolean
      subtitle:
        type: string
      title:
        type: string
      topic:
        type: string
    required:
    - authorname
    - content
    - subtitle
    - title
    type: object
  articles.Bookmark:
    properties:
      articles:
        items:
          type: integer
        type: array
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      userId:
        type: integer
    type: object
  articles.BookmarkInput:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
  articles.EditArticleInput:
    properties:
      content:
        type: string
      subtitle:
        type: string
      title:
        type: string
      topic:
        type: string
    required:
    - content
    - subtitle
    - title
    type: object
...
  articles.GetBookmarkObject:
    properties:
      articles:
        items:
          $ref: '#/definitions/main_router_bookmarks.Article'
        type: array
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      userId:
        type: integer
    type: object
...
  http.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  main_router_articles.Article:
    properties:
      authorName:
        type: string
      content:
        type: string
      draft:
        type: boolean
      id:
        type: integer
      likes:
        items:
          type: integer
        type: array
      subtitle:
        type: string
      title:
        type: string
      topic:
        type: string
      userId:
        type: integer
    type: object
  main_router_articles.DeletedResponse:
    properties:
      delete:
        example: all articles have been successfully deleted
        type: string
    type: object
  main_router_bookmarks.Article:
    properties:
      authorName:
        type: string
      content:
        type: string
      draft:
        type: boolean
      id:
        type: integer
      likes:
        items:
          type: integer
        type: array
      subtitle:
        type: string
      title:
        type: string
      topic:
        type: string
      userId:
        type: integer
    type: object
  main_router_bookmarks.DeletedResponse:
    properties:
      delete:
        example: all articles have been successfully deleted
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /articles:
    delete:
      consumes:
      - application/json
      description: Delete all the connected user articles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main_router_articles.DeletedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
    get:
      consumes:
      - application/json
      description: Retrieve all articles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main_router_articles.Article'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Add an article
      parameters:
      - description: Params to create an article
        in: body
        name: ArticleInput
        required: true
        schema:
          $ref: '#/definitions/articles.ArticleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main_router_articles.Article'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
    put:
      consumes:
      - application/json
      description: Edit an article
      parameters:
      - description: Params to edit an article
        in: body
        name: EditArticleInput
        required: true
        schema:
          $ref: '#/definitions/articles.EditArticleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main_router_articles.Article'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - articles
...
  /bookmarks:
    delete:
      consumes:
      - application/json
      description: Delete the connected user bookmarks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main_router_bookmarks.DeletedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - bookmarks
    get:
      consumes:
      - application/json
      description: Retrieve the connected user bookmarks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/articles.GetBookmarkObject'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - bookmarks
    post:
      consumes:
      - application/json
      description: Add a bookmark
      parameters:
      - description: Params to create a bookmark
        in: body
        name: BookmarkInput
        required: true
        schema:
          $ref: '#/definitions/articles.BookmarkInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/articles.Bookmark'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - bookmarks
  /bookmarks/:id:
    delete:
      consumes:
      - application/json
      description: Delete the connected user bookmark
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main_router_bookmarks.DeletedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - bookmarks
    get:
      consumes:
      - application/json
      description: Retrieve the connected user bookmark
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/articles.GetBookmarkObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - bookmarks
    put:
      consumes:
      - application/json
      description: Edit a bookmark
      parameters:
      - description: Params to edit a bookmark
        in: body
        name: BookmarkInput
        required: true
        schema:
          $ref: '#/definitions/articles.BookmarkInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/articles.Bookmark'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - bookmarks
  /bookmarks/:id/articles:
    delete:
      consumes:
      - application/json
      description: Remove all articles of a bookmark
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/articles.Bookmark'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - bookmarks
    get:
      consumes:
      - application/json
      description: Retrieve the articles of a given bookmark
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main_router_bookmarks.Article'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - bookmarks
  /bookmarks/:id/articles/:id-article:
    post:
      consumes:
      - application/json
      description: Add an article in a bookmark
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/articles.Bookmark'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - bookmarks
  /bookmarks/:id/articles/id-article:
    delete:
      consumes:
      - application/json
      description: Remove an article of a bookmark
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/articles.Bookmark'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.HTTPError'
      tags:
      - bookmarks
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"